Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
bcrypt
python-multipart
pymysql
redis
cryptography

# Add Kafka library
kafka-python==2.0.2
